Set the main window title properly, instead of abusing argv[0]
gtk_init()'s argv parameter is meant to be hooked up to the program's
command-line argument list, with the first arg being the program name
itself. GTK uses this first arg as the default title for the main window,
but strips all but the last path component so that running a GTK app as
"/usr/bin/myapp --foo" uses "myapp" as the default title. This mechanism
is thus unsuitable for implementing the :title adverb which is supposed to
be able to set arbitrary titles:
    GTK::Simple::App(title => "Enable / Disable")

This patch passes the program name as argv[0] like GTK expects, and then
explicitly sets the main window title using gtk_window_set_title() if the
:title adverb is specified.

This fixes #2