Deal with not being able to require 'Cairo' better
Previous the 'gtk_simple_use_cairo' would attempt to require Cairo
and would fail silently.  In turn the 'add-draw-handler' would then
attempt to to use values derived from the module by type lookup amd
pass them to a native routine, resulting in an exception in the
native code and subsequently a confusing MoarVM panic when they are
undefined.

This takes a belt and braces approach adding a Bool return to the
gtk_simple_use_cairo so user code can check and act appropriately
if Cairo can't be loaded, and the add-draw-handler will also
throw an exception if called without the module loaded.

Fixes #85