Add :subs option, export subroutine aliases for widget constructors
Now when you do:

  use GTK::Simple :subs;

you can subsequently call the constructor of any widget via a simplified
version of it's name:

  my $button = button(label => "Shorter");
  my $list-box = list-box;
  my $file-chooser-button = file-chooser-button(label => "File?");

The pattern is a simple conversion from camel case to kebab case.
Delicious!

This also fixes the issue of needing to `use GTK::Simple::App;` on a
separate line. Now it is available immediately.