Add a resource and a way to access it
In order to have a correct example for the new `GTK::Simple::Image`,
we need to use the `%?RESOURCES` of the `GTK::Simple` distribution.

This is exposed via:

  GTK::Simple::Distro::Resources::gtk-simple-resources

This is declared as a term so that once imported,
you can access the returned hash easily:

  gtk-simple-resources<camelia-logo.png>.slurp(:bin, :close)

This will hopefully be useful for users to some degree, but it
is primarily added for examples to access.