Add distro resources / handle files correctly in 22-image.raku (#113)
* Add GTK::Simple::Image

This requires a few new GDK related subs. I have placed them within
GTK::Simple::Raw instead of GTK::Simple::GDK because otherwise
we would need to import the latter into the former. Since these
new routines are only used by GTK::Simple::Image, it seems fine
to group them in Raw where they are imported via the :image import
set.

* Add necessary means for supporting MacOS

* Add a resource and a way to access it

In order to have a correct example for the new `GTK::Simple::Image`,
we need to use the `%?RESOURCES` of the `GTK::Simple` distribution.

This is exposed via:

  GTK::Simple::Distro::Resources::gtk-simple-resources

This is declared as a term so that once imported,
you can access the returned hash easily:

  gtk-simple-resources<camelia-logo.png>.slurp(:bin, :close)

This will hopefully be useful for users to some degree, but it
is primarily added for examples to access.

* Add MacOS::NativeLib dependency

* Only require MacOS::NativeLib on macOS

* Bump minor version to 0.3.2

* Mention Image in the readme

* Fix sub export test